requirements.txt
# Music Recommendations System - Dependencies

python-dotenv==1.2.1
requests==2.31.0
polars==1.35.2
joblib==1.3.2

catboost==1.2.2
fastapi==0.104.1
implicit==0.7.2
jupyterlab
lightfm==1.17
pandas==2.1.1
pyarrow==13.0.0
requests==2.31.0
scikit-learn==1.3.2
scikit-surprise==1.1.3
seaborn==0.12.1
uvicorn==0.24.0.post1

python-json-logger==4.0.0

